pyproject.toml
[project]
name = "support-ops-env"
version = "1.0.0"
description = "Customer support automation environment for AI agents"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic>=2.0",
    "openai",
    "python-dotenv",
]

[project.scripts]
server = "server.app:main"